+++ packages/moonstone/ExpandableItem/ExpandableItem.js
@@ -271,9 +271,9 @@ const ExpandableItemBase = kind({
 	},
 
 	computed: {
-		className: ({disabled, open, showLabel, styler}) => (styler.append({
+		className: ({disabled, label, noneText, open, showLabel, styler}) => (styler.append({
 			open: open && !disabled,
-			autoLabel: showLabel === 'auto'
+			autoLabel: showLabel === 'auto' && (label || noneText)
 		})),
 		label: ({label, noneText}) => (label || noneText),
 		labeledItemClassName: ({showLabel, styler}) => (styler.join(css.labeledItem, css[showLabel])),